// src/app/models/property.ts
export interface Property {
    Title: string;
    Description: string;
    Status: string;
    Type: string;
    RoomsNumber: number;
    Price: number;
    Area: number;
    Governorate: string;
    City: string;
    Age: string;
    BathroomNumber: number;
    Name: string;
    UserName: string;
    Email: string;
    Phone: string;
}